use vite add test sample
