+++ package.json
@@ -4,11 +4,10 @@
 	"description": "Sync TickTick tasks to Obsidian, and Obsidian tasks to TickTick",
 	"main": "main.js",
 	"scripts": {
-		"prepare": "npm run build-without-tsc",
-		"dev": "node esbuild.config.mjs",
-		"build": "tsc -noEmit -skipLibCheck && node esbuild.config.mjs production",
-		"build-without-tsc": "node esbuild.config.mjs production",
-		"version": "node version-bump.mjs && git add manifest.json versions.json"
+		"check": "tsc --noEmit",
+		"dev": " VITE_ENV=dev vite build --watch",
+		"build": "vite build",
+		"test": "vitest"
 	},
 	"keywords": [
 		"obsidian",
@@ -20,12 +19,6 @@
 	],
 	"author": "thesamim",
 	"license": "GNU GPLv3",
-	"dependencies": {
-		"@popperjs/core": "^2.11.6",
-		"bson-objectid": "^2.0.4",
-		"dependency-tree": "^11.0.1",
-		"ua-parser-js": "^1.0.39"
-	},
 	"repository": {
 		"type": "git",
 		"url": "https://github.com/thesamim/TickTickSync.git"
@@ -33,15 +26,28 @@
 	"bugs": {
 		"url": "https://github.com/thesamim/TickTickSync/issues"
 	},
+	"dependencies": {
+		"@popperjs/core": "^2.11.6",
+		"bson-objectid": "^2.0.4",
+		"dependency-tree": "^11.0.1",
+		"ua-parser-js": "^1.0.39"
+	},
 	"devDependencies": {
 		"@types/node": "^22.9.0",
+		"@types/ua-parser-js": "^0.7.39",
+		"@testing-library/jest-dom": "^6.4.2",
+		"@rollup/plugin-replace": "^5.0.5",
 		"@typescript-eslint/eslint-plugin": "^8.14.0",
 		"@typescript-eslint/parser": "^8.14.0",
 		"builtin-modules": "^4.0.0",
-		"esbuild": "^0.24.0",
-		"esbuild-plugin-copy": "^2.0.1",
+		"jsdom": "^24.0.0",
 		"obsidian": "latest",
 		"tslib": "^2.8.1",
-		"typescript": "^5.6.3"
+		"typescript": "^5.6.3",
+		"vite": "^5.1.0",
+		"vite-plugin-static-copy": "^1.0.1",
+		"vite-tsconfig-paths": "^4.3.2",
+		"vitest": "^1.2.2"
+
 	}
 }
