+++ src/modals/DefaultProjectModal.ts
@@ -1,5 +1,5 @@
 import { App, Modal, Setting } from "obsidian";
-import TickTickSync from "../../main"
+import TickTickSync from "@/main"
 
 
 interface MyProject {
